+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Releases before [0.4.0] have moved to [`CHANGELOG-archive.md`](CHANGELOG-archive
 
 ## [Unreleased]
 
+## [1.23.0] - 2026-09-25
+
 ### Removed
 
 - **The Symfony bundle's `prefix` option and